fix(inquirer-gui): apply line-height patch to siblings
