+++ packages/folder-browser-plugin/src/packages/QuestionFolderBrowser.vue
@@ -12,10 +12,15 @@
 </template>
 
 <script>
+import { mountLineHeightPatch } from "../../../inquirer-gui/src/utils";
+
 export default {
   name: "QuestionFolderBrowser",
   props: {
     question: Object,
+  },
+  mounted() {
+    mountLineHeightPatch(this.$refs.path);
   },
   data: () => ({
     path: "/home/",
