+++ packages/inquirer-gui/src/utils.js
@@ -1,3 +1,23 @@
+// @vscode-elements/elements@1.11.0 hardcodes line-height: 18px on the inner
+// <input> with no CSS custom property hook. Remove this patch when the library
+// exposes one (e.g. --vscode-input-line-height).
+function applyLineHeightPatch(el) {
+  if (el && el.shadowRoot && !el.shadowRoot.querySelector("style.line-height-patch")) {
+    const style = document.createElement("style");
+    style.className = "line-height-patch";
+    style.textContent = "input { line-height: normal !important; }";
+    el.shadowRoot.appendChild(style);
+  }
+}
+
+export function mountLineHeightPatch(el) {
+  if (el && typeof el.updateComplete !== "undefined") {
+    el.updateComplete.then(() => applyLineHeightPatch(el));
+  } else {
+    applyLineHeightPatch(el); // fallback for non-Lit environments; usually a no-op in tests
+  }
+}
+
 export default {
   // roughly based on underscore.js:
   //   https://github.com/jashkenas/underscore/blob/1.9.2/underscore.js#L887
