+++ packages/login-plugin/src/packages/QuestionLogin.vue
@@ -19,16 +19,15 @@
 </template>
 
 <script>
-// TODO: separate login from answer
-//   @input should fire an answerChanged event
-//   icon@click should fire a custom doLogin event
-// This enables validating the password upon typing
-// It also enables providing specific feedback when actual login fails/succeeds
+import { mountLineHeightPatch } from "../../../inquirer-gui/src/utils";
 
 export default {
   name: "QuestionLogin",
   props: {
     question: Object,
+  },
+  mounted() {
+    mountLineHeightPatch(this.$el);
   },
   methods: {
     afterLogin() {
